backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
black==25.1.0
blinker==1.9.0
boto3==1.40.5
botocore==1.40.5
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
click==8.2.1
# cloud-init==25.1.4
colorama==0.4.4
# command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==45.0.6
# dbus-python==1.2.18
# distro==1.7.0
# distro-info==1.1+ubuntu0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.110.1
flake8==7.3.0
Flask==3.1.1
frozenlist==1.7.0
# gunicorn==23.0.0
h11==0.16.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
# jeepney==0.7.1
Jinja2==3.1.6
jmespath==1.0.1
jq==1.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
# launchpadlib==1.10.16
lazr.restfulclient==0.14.4
# lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mercadopago==2.3.0
# meson==0.61.2
more-itertools==8.10.0
motor==3.3.1
multidict==6.6.3
mypy==1.17.1
mypy_extensions==1.1.0
# netifaces==0.11.0
# nftables==0.1
numpy==2.2.6
oauthlib==3.2.0
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pathspec==0.12.1
# pexpect==4.8.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.2
# ptyprocess==0.7.0
pyasn1
pyasn1-modules
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
# PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.10.1
pymongo==4.5.0
pyOpenSSL==25.1.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
pytest==8.4.1
# python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
# python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.1
python-http-client==3.3.7
python-jose==3.5.0
python-magic==0.4.24
python-multipart==0.0.20
pytz==2022.1
# PyYAML==5.4.1
PyYAML>=6.0
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
s3transfer==0.13.1
# SecretStorage==3.3.1
sendgrid==6.11.0
service-identity==18.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
# sos==4.8.2
# ssh-import-id==5.11
starkbank-ecdsa==2.2.0
starlette==0.37.2
# systemd-python==234
tomli==2.2.1
Twisted==22.1.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
# ubuntu-pro-client==8001
# ufw==0.36.1
# unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.25.0
wadllib==1.3.6
Werkzeug==3.1.3
yarl==1.20.1
zipp==1.0.0
zope.interface==5.4.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.4.0
